refactor: rename conponents -> components
