+++ src/components/PropertyItem/PropertyItem.tsx
