+++ src/components/PropertyItem/index.ts
