+++ src/components/WeatherProperties/WeatherProperties.tsx
