+++ src/components/WeatherProperties/index.ts
